Make the modification methods keep the cache up-to-date
There is a cache in the module. However it's not updated by the
`create_record()`, `update_record()` and `delete_record()` methods. As
a result, the `sync_dns_from_my_ip()` method updates the record on
Cloudflare, but not in the `dns_records` list. Therefore, the condition
`record['content'] != ip_address` is always true.